Port the ClockworkPi trackball driver to QMK.
This is a port of the Arduino source code in
https://github.com/clockworkpi/uconsole.

I've tried to keep this as true to the original implementation as
possible so the performance and behavior are comparable.

Note, however, I have added one trivial new feature: horizontal
scroll-wheel support! :D

Additionally, add missing header while bumping up the QMK version.

Co-Authored-By: perillamint <perillamint@silicon.moe>